Move Clock classes to SDK\Common\Time namespace
