+++ src/SDK/Common/Time/ClockInterface.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace OpenTelemetry\SDK;
+namespace OpenTelemetry\SDK\Common\Time;
 
 interface ClockInterface
 {
@@ -12,7 +12,7 @@ interface ClockInterface
     /**
      * Returns the current epoch wall-clock timestamp in nanoseconds.
      * This timestamp should _ONLY_ be used to compute a current time.
-     * Use {@see \OpenTelemetry\SDK\AbstractClock::nanoTime} for calculating
+     * Use {@see \OpenTelemetry\SDK\Common\Time\AbstractClock::nanoTime} for calculating
      * durations.
      */
     public function now(): int;
