+++ tests/Unit/SDK/Common/Time/SystemClockTest.php
@@ -2,20 +2,16 @@
 
 declare(strict_types=1);
 
-namespace OpenTelemetry\Tests\Unit\SDK;
+namespace OpenTelemetry\Tests\Unit\SDK\Common\Time;
 
-use OpenTelemetry\SDK\SystemClock;
+use OpenTelemetry\SDK\Common\Time\SystemClock;
 use PHPUnit\Framework\TestCase;
 
 /**
- * @covers OpenTelemetry\SDK\SystemClock
+ * @covers \OpenTelemetry\SDK\Common\Time\SystemClock
  */
 class SystemClockTest extends TestCase
 {
-    /**
-     * @runInSeparateProcess
-     * @preserveGlobalState false
-     */
     public function test_get_instance_always_returns_same_clock(): void
     {
         $clock = SystemClock::getInstance();
