+++ tests/Unit/SDK/Util/SpanData.php
@@ -7,10 +7,10 @@ namespace OpenTelemetry\Tests\Unit\SDK\Util;
 use function count;
 use function max;
 use OpenTelemetry\API\Trace as API;
-use OpenTelemetry\SDK\AbstractClock;
 use OpenTelemetry\SDK\Common\Attribute\Attributes;
 use OpenTelemetry\SDK\Common\Attribute\AttributesInterface;
 use OpenTelemetry\SDK\Common\Instrumentation\InstrumentationLibrary;
+use OpenTelemetry\SDK\Common\Time\AbstractClock;
 use OpenTelemetry\SDK\Resource\ResourceInfo;
 use OpenTelemetry\SDK\Trace as SDK;
 use OpenTelemetry\SDK\Trace\EventInterface;
